Creation of the app dir, moved home page and about page
