+++ components/BrandDisclaimer.tsx
@@ -1,0 +1,12 @@
+export default function BrandDisclaimer() {
+  return (
+    <div className='pt-5'>
+      <p className='small'>
+        Kentobeans7, Kentobot, and Kentobeans.live is not affiliated with any of
+        brands included on this page, and inclusion of a name and/or logo is not
+        meant to imply or suggest any endorsement, sponsorship, or involvement,
+        direct or indirect.
+      </p>
+    </div>
+  );
+}
