+++ package.json
@@ -17,6 +17,7 @@
     "@fortawesome/react-fontawesome": "^0.2.0",
     "@vercel/analytics": "^0.1.11",
     "axios": "^1.3.4",
+    "bootstrap": "^5.3.2",
     "cookies-next": "^2.1.1",
     "i": "^0.3.7",
     "moment": "^2.29.4",
