+++ styles/globals.css
@@ -1,3 +1,6 @@
+@import "~bootstrap/dist/css/bootstrap.min.css";
+
+
 /*
 DARK BLUE - #0E1336;
 LIGHT BLUE - #4376BB;
